Refactor: Organize the project in directories

+++ philo/includes/philo.h
@@ -6,7 +6,7 @@
 /*   By: victofer <victofer@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/15 18:48:21 by victofer          #+#    #+#             */
-/*   Updated: 2023/03/29 18:18:47 by victofer         ###   ########.fr       */
+/*   Updated: 2023/03/29 18:22:37 by victofer         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ long long	ft_atoi(const char *str);
 t_table		*init_table(t_table *table, int argc, char **argv);
 t_philo		**init_philos(t_table *table);
 void		asign_forks(t_philo *philos);
-unsigned long	get_timestamp_ms(void);
+long long	get_timestamp_ms(void);
 
 //	C H E C K S
