+++ philo/src/time.c
@@ -6,21 +6,21 @@
 /*   By: victofer <victofer@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/20 18:32:34 by victofer          #+#    #+#             */
-/*   Updated: 2023/03/29 18:19:08 by victofer         ###   ########.fr       */
+/*   Updated: 2023/03/30 10:16:25 by victofer         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "philo.h"
+#include "../includes/philo.h"
 
 /* 
  * get_timestamp_ms
  * ----------------------------
  *   Returns the current timestamp converted to miliseconds.
  */
-unsigned long	get_timestamp_ms(void)
+long long	get_timestamp_ms(void)
 {
 	struct timeval	cur;
-	unsigned long		time;
+	long long		time;
 
 	gettimeofday(&cur, NULL);
 	time = (cur.tv_sec) * 1000 + (cur.tv_sec) / 1000;
